requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
insightface==0.7.3
onnxruntime==1.16.0
opencv-python==4.8.1.78
numpy==1.24.4
pillow==10.1.0
python-multipart==0.0.6
requests==2.31.0
scikit-learn==1.3.2
tqdm==4.66.1
